Dev: 추가 기능 및 배포 작성

+++ components/post/postService.js
@@ -1,28 +1,42 @@
-const { Post } = require("../../models/index");
+const { Post, User } = require("../../models/index");
 const crypto = require("crypto");
+
+
 
 /**
  * 모든 게시글을 정렬해서 찾아줌
  */
-const allPostsWithOrdering = async () => {
+const allPostsWithOrdering = async (offset,limit) => {
   const all = await Post.findAll({
-    attributes: ["title", "content", "created_at"],
+    attributes: ["postId","title", "content", "weather","created_at"],
     order: [["created_at", "DESC"]],
+    offset: offset,
+    limit: limit,
   }).catch((err) => {
     throw new Error(err);
   });
   return all;
 };
 
-const findWeather = async (userId) => {
-  const userLocation = await Post.findOne({ userId: userId }).catch((err) => {
+/**
+ * 사용자 id에 해당하는 장소 조회
+ * @param {number} userId 
+ * @returns {StringConstructor}
+ */
+const findLocation = async (userId) => {
+  const userLocation = await User.findOne({ userId: userId }).catch((err) => {
     throw new Error(err);
   });
   const location = userLocation.dataValues.location;
-  // 날씨 찾아오기
-  return weatherState;
+  return location;
 };
 
+/**
+ * 게시글 수정시 등록된 비밀번호와 같은지 조회
+ * @param {number} postId 
+ * @param {string} hashPassword 
+ * @returns {boolean}
+ */
 const userAuth = async (postId, hashPassword) => {
   const findPostPassword = await Post.findOne({
     postId: postId,
@@ -34,7 +48,15 @@ const userAuth = async (postId, hashPassword) => {
   }
 };
 
+/**
+ * 게시글에서 등록한 비밀번호가 조건에 맞게 작성되었는지 확인
+ * @param {string} password 
+ * @returns {boolean}
+ */
 const newPasswordAuth = (password) => {
+  if(isNaN(parseInt(password))){
+    return false;
+  }
   if (password.length > 6 && isNaN(password)) {
     return true;
   } else {
@@ -42,7 +64,16 @@ const newPasswordAuth = (password) => {
   }
 };
 
-const postPosted = async (title, content, password, userId, weather) => {
+/**
+ *  게시글 등록
+ * @param {string} title 
+ * @param {string} content 
+ * @param {string} password 
+ * @param {number} userId 
+ * @param {string} weather 
+ * @returns 
+ */
+const postPosted = async (title, content, password, userId,weather) => {
   const hashPassword = await crypto
     .createHash("sha256")
     .update(password)
@@ -59,6 +90,16 @@ const postPosted = async (title, content, password, userId, weather) => {
   return create;
 };
 
+/**
+ * 게시글 수정
+ * @param {number} postId 
+ * @param {string} title 
+ * @param {string} content 
+ * @param {string} password 
+ * @param {number} userId 
+ * @param {string} weather 
+ * @returns 
+ */
 const postUpdeted = async (
   postId,
   title,
@@ -93,21 +134,37 @@ const postUpdeted = async (
   }
 };
 
-const postDeleted = async (postId) => {
-  const destroyResult = await Post.destroy({
-    where: { postId: postId },
-  }).catch((err) => {
-    throw new Error(err);
-  });
+/**
+ * 게시글 삭제
+ * @param {number} postId 
+ * @param {string} password 
+ * @returns 
+ */
+const postDeleted = async (postId,password) => {
+  const hashPassword = crypto
+    .createHash("sha256")
+    .update(password)
+    .digest("base64");
+  const isPasswordTrue = await userAuth(userId, hashPassword);
+  if (isPasswordTrue) {
+    const destroyResult = await Post.destroy({
+      where: { postId: postId },
+    }).catch((err) => {
+      throw new Error(err);
+    });
+    return destroyResult;
+  } else {
+    throw new Error('올바른 비밀번호를 입력하세요');
+  }
 
-  return destroyResult;
+  
 };
 
 module.exports = {
   postDeleted,
   postPosted,
   postUpdeted,
-  findWeather,
+  findLocation,
   newPasswordAuth,
   allPostsWithOrdering,
 };
